Fixed a lot of error catching in REST/PROXY and added password to SOAP

+++ src/sd/tp1/svr/SOAP/SharedGalleryServerSOAP.java
@@ -3,7 +3,9 @@ package sd.tp1.svr.SOAP;
 import sd.tp1.svr.SharedGalleryFileSystemUtilities;
 import sd.tp1.svr.SharedGalleryClientDiscovery;
 
+import java.io.BufferedReader;
 import java.io.File;
+import java.io.InputStreamReader;
 import java.net.*;
 import java.util.*;
 
@@ -19,6 +21,7 @@ import javax.xml.ws.Endpoint;
 public class SharedGalleryServerSOAP {
 
     private static File basePath = new File("./FileServerSOAP");
+    private static String local_password;
 
     /**
      * The methods from this class act the same way as the ones from REQUEST interface
@@ -30,39 +33,59 @@ public class SharedGalleryServerSOAP {
     }
 
     @WebMethod
-    public List<String> getListOfPictures(String album) {
-        return SharedGalleryFileSystemUtilities.getPicturesFromDirectory(basePath, album);
+    public List<String> getListOfPictures(String album, String password) {
+        if(validate(password))
+            return SharedGalleryFileSystemUtilities.getPicturesFromDirectory(basePath, album);
+        return null;
     }
 
     @WebMethod
-    public byte [] getPictureData(String album, String picture) {
-        return SharedGalleryFileSystemUtilities.getDataFromPicture(basePath, album, picture);
+    public byte [] getPictureData(String album, String picture, String password) {
+        if(validate(password))
+            return SharedGalleryFileSystemUtilities.getDataFromPicture(basePath, album, picture);
+        return null;
     }
 
     @WebMethod
-    public String createAlbum(String album) {
-        return SharedGalleryFileSystemUtilities.createDirectory(basePath, album);
+    public String createAlbum(String album, String password) {
+        if(validate(password))
+            return SharedGalleryFileSystemUtilities.createDirectory(basePath, album);
+        return null;
     }
 
     @WebMethod
-    public Boolean deleteAlbum(String album){
-        return SharedGalleryFileSystemUtilities.deleteDirectory(basePath, album);
+    public Boolean deleteAlbum(String album, String password){
+        if(validate(password))
+            return SharedGalleryFileSystemUtilities.deleteDirectory(basePath, album);
+        return false;
     }
 
     @WebMethod
-    public String uploadPicture(String album, String picture, byte [] data) {
-        return SharedGalleryFileSystemUtilities.createPicture(basePath, album, picture, data);
+    public String uploadPicture(String album, String picture, byte [] data, String password) {
+        if(validate(password))
+            return SharedGalleryFileSystemUtilities.createPicture(basePath, album, picture, data);
+        return null;
     }
 
     @WebMethod
-    public Boolean deletePicture(String album, String picture) {
-        return SharedGalleryFileSystemUtilities.deletePicture(basePath, album, picture);
+    public Boolean deletePicture(String album, String picture, String password) {
+        if(validate(password))
+            return SharedGalleryFileSystemUtilities.deletePicture(basePath, album, picture);
+        return false;
+    }
+
+    private static boolean validate(String password) {
+        return password.equalsIgnoreCase(local_password);
     }
 
     public static void main(String args[]) throws Exception {
 
-        if(!basePath.exists())
-            basePath.mkdir();
+        BufferedReader reader = new BufferedReader(new InputStreamReader(System.in));
+
+        System.out.println("LOCAL PASSWORD: ");
+        local_password = reader.readLine();
+
+        basePath.mkdir();
 
         String address_s = "http://" + InetAddress.getLocalHost().getHostAddress() + ":8080/FileServerSOAP";
         Endpoint.publish(address_s, new SharedGalleryServerSOAP());
@@ -70,8 +93,7 @@ public class SharedGalleryServerSOAP {
         System.err.println("SharedGalleryServerSOAP: Started @ " + address_s);
 
         // Receives
-        Thread r = new Thread(new SharedGalleryClientDiscovery(address_s));
-        r.start();
+        new Thread(new SharedGalleryClientDiscovery(address_s)).start();
 
     }
